fixup! KTOR-7299 Fix "Fail to select descriptor…" error on descriptor close
